Program now runs from program.cs Updated plane wireframe to segment off the cabin space Added more outside lights to the back of the plane Added in a proper ID scheme for the lights to be identified (different from the display names) I also tried to save lines by removing unecessary ones
