+++ Avionic-Lighting-System/Program.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Windows.Forms;
+
+namespace AircraftLightsGUI {
+    internal static class Program {
+        [STAThread]
+        static void Main(){
+            Application.EnableVisualStyles();
+            Application.SetCompatibleTextRenderingDefault(false);
+            Application.Run(new MainForm());
+        }
+    }
+}
