+++ Avionic-Lighting-System/bin/Debug/net9.0-windows/Avionic-Lighting-System.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v9.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v9.0": {
+      "Avionic-Lighting-System/1.0.0": {
+        "runtime": {
+          "Avionic-Lighting-System.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "Avionic-Lighting-System/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
