+++ Avionic-Lighting-System/bin/Debug/net9.0-windows/Avionic-Lighting-System.runtimeconfig.json
