+++ Avionics-Lighting-Interface.sln
@@ -2,7 +2,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Avionics-Lighting-Interface", "Avionics-Lighting-Interface.csproj", "{3806438E-9F8A-8970-9C35-459DB98001E1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Avionic-Lighting-System", "Avionic-Lighting-System\Avionic-Lighting-System.csproj", "{A297409E-A7F6-34AD-CDB5-7A304AA941D3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,15 +10,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3806438E-9F8A-8970-9C35-459DB98001E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3806438E-9F8A-8970-9C35-459DB98001E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3806438E-9F8A-8970-9C35-459DB98001E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3806438E-9F8A-8970-9C35-459DB98001E1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A297409E-A7F6-34AD-CDB5-7A304AA941D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A297409E-A7F6-34AD-CDB5-7A304AA941D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A297409E-A7F6-34AD-CDB5-7A304AA941D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A297409E-A7F6-34AD-CDB5-7A304AA941D3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6C6A9A40-92E8-4D69-B210-857575D37BEC}
+		SolutionGuid = {FD4E0061-0FD5-4D52-9175-588211111719}
 	EndGlobalSection
 EndGlobal
