--- README.md
@@ -1,1 +1,0 @@
-# Avionics-Lighting-Interface
