--- bin/Debug/net9.0-windows/Avionics-Lighting-Interface.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v9.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v9.0": {
-      "Avionics-Lighting-Interface/1.0.0": {
-        "runtime": {
-          "Avionics-Lighting-Interface.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "Avionics-Lighting-Interface/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
